ffprobe/avtextformat: Move formatters to avutil
